added decimal option for level

+++ db/migrate/20170227163051_create_events.rb
@@ -5,7 +5,7 @@ class CreateEvents < ActiveRecord::Migration[5.0]
             t.string :posted_by
             t.datetime :created_at
             t.string :location
-            t.integer :level
+            t.decimal :level
             t.text :description
 
     end
